fixed windows linking and warning

+++ compiler/code_output.cpp
@@ -18,7 +18,7 @@
  * Various methods to return strings which are used very often
  */
 
-const char* get_reg_type(uint8_t req_type)
+const char* get_reg_type(int req_type)
 {
     switch(req_type)
     {
